Some test refactor
Do better pkg names next time mike.
